third_party/blink/web_tests/flag-specific/highdpi/virtual/css-calc-infinity-and-nan/external/wpt/css/css-values/calc-infinity-nan-computed-expected.txt

This is a testharness.js-based test.
FAIL Property width value 'calc(NaN * 1px)' assert_greater_than_equal: calc(NaN * 1px) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(infinity * 1px)' assert_greater_than_equal: calc(infinity * 1px) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(infinity * 1cm)' assert_greater_than_equal: calc(infinity * 1cm) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(NaN * 1rem)' assert_greater_than_equal: calc(NaN * 1rem) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(infinity * 1px - infinity * 1%)' assert_greater_than_equal: calc(infinity * 1px - infinity * 1%) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(infinity * 1px + infinity * 1%)' assert_greater_than_equal: calc(infinity * 1px + infinity * 1%) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(min(NaN * 1px, infinity * 1px) + max(infinity * 1px, -infinity * 1px))' assert_greater_than_equal: calc(min(NaN * 1px, infinity * 1px) + max(infinity * 1px, -infinity * 1px)) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(infinity * 1px - max(infinity * 1%, 0%))' assert_greater_than_equal: calc(infinity * 1px - max(infinity * 1%, 0%)) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property width value 'calc(max(infinity * 1px, 10px))' assert_greater_than_equal: calc(max(infinity * 1px, 10px)) expected a number greater than or equal to 33554400 but got 22369600
FAIL Property margin-left value 'calc(-infinity * 1px)' assert_less_than_equal: calc(-infinity * 1px) expected a number less than or equal to -33554400 but got -22369600
FAIL Property margin-left value 'calc(min(1px, -infinity * 1%))' assert_less_than_equal: calc(min(1px, -infinity * 1%)) expected a number less than or equal to -33554400 but got -22369600
FAIL Property margin-left value 'calc(-infinity * 1%)' assert_less_than_equal: calc(-infinity * 1%) expected a number less than or equal to -33554400 but got -22369600
FAIL Property margin-left value 'calc(max(10000px, 0px) + min(-infinity * 1px, infinity * 1px))' assert_less_than_equal: calc(max(10000px, 0px) + min(-infinity * 1px, infinity * 1px)) expected a number less than or equal to -33554400 but got -22369600
FAIL Property margin-left value 'calc(-infinity * 1px - infinity * 1px)' assert_less_than_equal: calc(-infinity * 1px - infinity * 1px) expected a number less than or equal to -33554400 but got -22369600
FAIL Property margin-left value 'calc(min(-infinity * 1px, 10px))' assert_less_than_equal: calc(min(-infinity * 1px, 10px)) expected a number less than or equal to -33554400 but got -22369600
PASS Property animation-duration value 'calc(NaN * 1s)'
PASS Property animation-duration value 'calc(infinity * 1s)'
PASS Property animation-duration value 'calc(1 / 0 * 1s)'
PASS Property animation-duration value 'calc(max(infinity * 1s, 10s)'
PASS Property transition-delay value 'calc(-infinity* 1s)'
PASS Property transition-delay value 'calc(max(10000s, 0s) + min(-infinity * 1s, infinity * 1s))'
PASS Property transition-delay value 'calc(min(-infinity * 1s, 10s))'
Harness: the test ran to completion.
